Separated geo-spatial inputs to its own file
