+++ gaia/geo/__init__.py
@@ -1,4 +1,4 @@
 from gaia.geo.gdal_functions import *
-from gaia.geo.gaia_process import *
+from gaia.gaia_process import *
 from gaia.geo.processes_raster import *
 from gaia.geo.processes_vector import *
