+++ gaia/geo/processes_raster.py
@@ -18,9 +18,9 @@
 ###############################################################################
 import logging
 import gaia.formats as formats
-from gaia.geo.gaia_process import GaiaProcess
+from gaia.gaia_process import GaiaProcess
 from gaia.geo.gdal_functions import gdal_calc, gdal_clip
-import gaia.inputs
+from gaia.geo.geo_inputs import RasterFileIO
 
 
 logger = logging.getLogger('gaia.geo')
@@ -44,8 +44,7 @@ class SubsetProcess(GaiaProcess):
         """
         super(SubsetProcess, self).__init__(**kwargs)
         if not self.output:
-            self.output = gaia.inputs.RasterFileIO(name='result',
-                                                   uri=self.get_outpath())
+            self.output = RasterFileIO(name='result', uri=self.get_outpath())
         self.validate()
 
     def compute(self):
@@ -80,8 +79,7 @@ class RasterMathProcess(GaiaProcess):
         super(RasterMathProcess, self).__init__(**kwargs)
         self.calc = calc
         if not self.output:
-            self.output = gaia.inputs.RasterFileIO(name='result',
-                                                   uri=self.get_outpath())
+            self.output = RasterFileIO(name='result', uri=self.get_outpath())
         self.validate()
 
     def compute(self):
